app/requirements.txt
requests
qdrant-client
openai
anthropicpsycopg2-binary